backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
feedparser==6.0.11
apscheduler==3.10.4
slowapi==0.1.9
redis==5.0.3
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
